remove more unused code, and keep unused vars with side-effects but add an annotation so it won't cause eclipse warnings

+++ src/main/java/org/vngx/jsch/Buffer.java
@@ -392,7 +392,7 @@ public final class Buffer {
 	 */
 	public long getUInt() {
 		return  (((long) getShort() << 16) & 0xFFFF0000L) |	// First 16 bytes
-				(((long) getShort()      ) & 0x0000FFFFL);	// second 16 bytes
+				((       getShort()      ) & 0x0000FFFFL);	// second 16 bytes
 	}
 
 	/**
